Session 3 notes: prefer linked-list traversal via 'next' pointer
Following the 'next' pointer avoids problems when the memory isn't
contiguous.

The code as previously written worked because the elements were in
immediate ascending order. For example, if there was another variable
allocated on the stack in between Elements, the code would segfault:

    struct Element tail = { &j, NULL};
    int k = 5;
    struct Element head = { &i, &tail };

Similarly, for heap-allocated Elements, incrementing the pointer would
result in unexpected/incorrect behavior.